add drag and drop feature to turn an image into a label and container (#7553)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Adds ability to drag and drop an image onto the main window. Image is
turned into a label inside an AdjustableContainer. Image is copied to
profile-subdir/GUIDropImages.

#### Motivation for adding to Mudlet
Code is basically unchanged from Edru's input #1399 . I just packaged it
up as per discussion with Vadi. Also created a mpackage if we ever get
#7268 working.

#### Other info (issues closed, discussion etc)