Fix: crash on installing packages on Windows (#7571)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Fixed a crash when installing packages on Windows
#### Motivation for adding to Mudlet
Fix https://github.com/Mudlet/Mudlet/issues/7255
#### Other info (issues closed, discussion etc)
It boils down to the issue of the 'are we saving a profile?' check being
_before_ we have set 'we are saving a profile'. Now the check is
correctly set afterwards.

Longer version:
Installing a package necessitated saving the profile - however, in the
middle of the save profile operation, control was being returned back to
the running code. The next step in installing a package that already
existed was to uninstall it, which started the save profile operation
for the second time - while the first save profile was still ongoing!

This seems to have caused the issues. How, installing a package still
saves the profile, but control is not returned to the running code until
we have set the fact that we are now saving the profile. This means that
when the second save profile operation is attempted to be during package
uninstall, it doesn't go through, and we don't have two profile save
operations happening in parallel.

---------

Co-authored-by: Vadim Peretokin <vadi2@users.noreply.github.com>