Fix: Duplication of module contents after saving with module running on multiple loaded profiles (#7572)
#### Brief overview of PR changes/additions
Remove duplicate writes from reloadModule() function. The editing
profile writes to the module (in saveProfile function), and then
reloadModule uninstalls and reinstalls the module in all the other open
profiles. Writing the module xml again in the other open profiles causes
the duplication.

#### Motivation for adding to Mudlet
To fix the duplication bug in modules

#### Other info (issues closed, discussion etc)
Closes issue 7484
/claim #7484