Fix: show room symbols at smaller sizes (#7593)
#### Brief overview of PR changes/additions
Allow the font used to draw room symbols to be half the previous minimum
so that they show at even smaller sizes and thus are shown when the map
is zoomed out more.

#### Motivation for adding to Mudlet
Someone (@Kebap) was unhappy at the size at which symbols were not being
shown.

#### Other info (issues closed, discussion etc)
If accepted this will close #5320 and allow me to /claim #5320.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Kebap <kebap_spam@gmx.net>