Fix: Windows updater 'deletes' Mudlet (#7620)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Change the location of the newly downloaded installer from the same
folder that mudlet.exe to a different, temporary writable directory.
This is because Squirrel tries to delete the folder where mudlet.exe is
fully first, which sometimes is an issue because that's the same
location that the installer is running from.
#### Motivation for adding to Mudlet
Fix the important https://github.com/Mudlet/Mudlet/issues/7313 issue.
#### Other info (issues closed, discussion etc)

---------

Co-authored-by: Vadim Peretokin <vadi2@users.noreply.github.com>