Fix: ensure a script editor list item is always selected (#7663)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
In the Script Editor, ensure the root item is selected on startup. When
deleting an item, ensure the parent folder (or root) is selected. This
clearly shows which folder is being worked on and prevents a new item
being added to an unexpected bottom folder.

#### Motivation for adding to Mudlet
Prevents unexpected behavior, better user experience.  

#### Other info (issues closed, discussion etc)
[Peek 2025-01-15
05-20.webm](https://github.com/user-attachments/assets/de08cff3-e34c-4121-b27b-58b3a2354df2)

/closes #7632