Improve: make map label input box handle more than one line. (#7674)
#### Brief overview of PR changes/additions
Changes the text entry widget for the text content for that type of "Map
Label" from a `QLineEdit` to a `QPlainTextEdit`.

#### Motivation for adding to Mudlet
To make it possible to enter multi-line texts into text Map Labels. The
existing widget does not show/enable the native entry of line-feeds and
the only way to get them into such texts is to prepare them in an
external location and paste them in - and after doing so they look just
like spaces.

#### Other info (issues closed, discussion etc)
This the first (I hope) of some improvements I want to make to the map
label user-experience. I found out that this was needed whilst MUDding
on After the Plague MUD where, when map making I wanted to label area
exits with some extra details and to include textual information about
locations within the map.

I also discovered the `(bool) tabChangesFocus` property which several
text entry widgets possess which if set (and it is normally `false`)
allows the `<Tab>` key to be used to move the focus out of that widget;
when it is `false` it is used to enter *tabs* into the text itself. For
this widget I have set it to `true` - but I do wonder if there are other
widgets in the Mudlet application where this needs to be reviewed
because it does impact on the user experience particularly for
visually-impaired persons.

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>