fix: Enable connect and offline buttons when selecting currently open and then different profile (#7692)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
This enables the Connect and Offline buttons when changing selection
from a currently loaded profile over to another profile that has same
server address and port number.

#### Motivation for adding to Mudlet
I prioritize my time badly and sometimes get the wrong things off my
plate first. Though I'm thinking of this as a mental health break from
other tasks.

#### Other info (issues closed, discussion etc)
The changes in #7673 made it so the connection dialog will disable the
Connect and Offline buttons when selecting a profile that is already
open. But re-enabling the buttons when selecting another profile only
happens when other profile has different server address or port. So if
you have two profiles for same game, need to select an intermediate one
to enable the buttons currently.

Whenever you switch to a different profile that has a different server
address, the box gets updated which triggers `slot_updateUrl`, which in
turn calls `validateProfile()`. Similar deal with `slot_updatePort`. But
`slot_updateName` is not similarly called when the name box changes,
only when you are typing in the name box, so it only runs
`validateProfile()` when typing. I found that the difference between it
and the others is `::textChanged` vs `::textEdited`.