Fix reduce c++20warnings (#7638)
#### Brief overview of PR changes/additions
1. Adds explicit ‘this’ or ‘*this’ capture to lambdas where required
(not in static ones).
2. Rephrases some combinations of Qt `enum`s that whilst acceptable for
C++17 upset the later one. For the non-key related ones which are
provided a `int`s arguments to a method this is sufficient; For Qt 6 there are
some additional methods that overcome the "incompatibility" of combining
such different `enum`s but those have not been back ported to the Qt
5.15.8 I have - despite what https://bugreports.qt.io/browse/QTBUG-99948
says - so fixes for combining QKey and QKeyModifier have been left out of
this PR.
3. Removes a couple of unused variables.
4. Adds our `qsl(...)` wrapper about some raw C-string literals used in
a loop.
5. Rewrite part of the qmake project file so the logic and choices are
correct.

#### Motivation for adding to Mudlet
1. To eliminate the following type of warning when building with a C++20
compiler:
"warning: implicit capture of ‘this’ via ‘[=]’ is deprecated in C++20
[-Wdeprecated]"
2. To eliminate the following type of warning when building with a C++20
compiler:
* warning: bitwise operation between different enumeration types
‘QFont::Weight’ and ‘QFont::StyleHint’ is deprecated
[-Wdeprecated-enum-enum-conversion]
* warning: bitwise operation between different enumeration types
‘Qt::TextFlag’ and ‘Qt::AlignmentFlag’ is deprecated
[-Wdeprecated-enum-enum-conversion]
3. To eliminate the following type of warning when building with a C++20
compiler:
"warning: unused variable ‘pHost’ [-Wunused-variable]"
4. To eliminate the following type of warning when building with a C++20
compiler:
"warning: loop variable ‘file’ of type ‘const QString&’ binds to a
temporary constructed from type ‘const char* const’
[-Wrange-loop-construct]"
5. The previous logic was (incorrect)
  IF Qt Major version is less than 5
OR (if Qt Major version is less than 6 AND if Qt Minor version is less
than 12))
    THEN add `-std=c++20` to `QMAKE_CXXFLAGS`
    ELSE add `c++2a` to `CONFIG`
This is borked because we have already rejected Qt versions less than
5.14 so the logic will **always** end up in the **ELSE** case, and
whilst Qt 5.15 is documented as accepting `CONFIG += c++2a` (and
`c++2b`) the earliest Qt 6.x version describes `c++2a` as an obsolete
alias for `c++20` (in
https://doc.qt.io/qt-6.2/qmake-variable-reference.html).
  
#### Other info (issues closed, discussion etc)
3. The code in the `QFont` cases contains errors that have been copied
from bogus QFont creation code going back to the very first (well
second) commit in the git history. This contained calls of the form
`QFont font("Courier New", 10, QFont::Courier)`
however even the Qt 4.8 documentation does not list a constructor of
that form but instead has:
`QFont(const QString & family, int pointSize = -1, int weight = -1, bool
italic = false)`
the third argument could possibly be `QFont::Normal` (50) or
`QFont::Bold` (75) however the value of `QFont::Courier` is `2`
but it is for a completely different purpose, that of the font matching
strategy ("the font matcher prefers fixed pitch fonts.") but that is not
something that can be set as an argument to the font constructor!
  
As it happens the combination of `QFont::Bold | QFont::Serif |
QFont::PreferMatch | QFont::PreferAntialias` that was being used
numerically equals 75 + 2 + 32 + 128 = 227 - and the scale that Qt
actually uses only goes from 0 to 99!

5. This change makes the code match the comments!

Overall all the changes in this PR means that
https://github.com/Mudlet/Mudlet/pull/7613 is not required after all -
at least for Mudlet's own code - though there are still some warnings
from the edbee-lib sub-module.

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>